+++ src/queues/distributed_queue.js
@@ -58,8 +58,10 @@ class Queue {
         let consumeOptions = _.merge({}, this.config, options);
         let queue = consumeOptions.name || channel.__queue;
 
-        if (!channel)
-            channel = await this._getAndValidateConsumeChannel(queue, topic, options);
+        if (!channel) {
+            await this._validateConsumeChannel(queue);
+            channel = await this.channelManager.getConsumeChannel(topic, options);
+        }
 
         this.consumers.unshift({ channel, handler, topic, options, queue });
 
@@ -207,14 +209,12 @@ class Queue {
         });
     }
 
-    async _getAndValidateConsumeChannel(queue, topic, options) {
+    _validateConsumeChannel(queue) {
         if (this.consumers.length) {
             let sameQueueConsumer = this.consumers.find(x => x.queue === queue);
             if (sameQueueConsumer)
                 throw new Error('Multiple consumers registered to the same queue. If you meant to add bindings, use the "addTopic" method instead');
         }
-
-        return await this.channelManager.getConsumeChannel(topic, options);
     }
 }